Add more code isolation
