+++ test_runner/regress/test_basebackup.py
@@ -2,26 +2,10 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-import pytest
 from fixtures.utils import wait_until
 
 if TYPE_CHECKING:
-    from fixtures.neon_fixtures import NeonEnv, NeonEnvBuilder
-
-
-def test_basebackup_error(neon_simple_env: NeonEnv):
-    """
-    Test error handling, if the 'basebackup' command fails in the middle
-    of building the tar archive.
-    """
-    env = neon_simple_env
-    pageserver_http = env.pageserver.http_client()
-
-    # Introduce failpoint
-    pageserver_http.configure_failpoints(("basebackup-before-control-file", "return"))
-
-    with pytest.raises(Exception, match="basebackup-before-control-file"):
-        env.endpoints.create_start("main")
+    from fixtures.neon_fixtures import NeonEnvBuilder
 
 
 def test_basebackup_cache(neon_env_builder: NeonEnvBuilder):
